chore : 동적 라우팅 수정
