+++ service/gateway/gateway_domain/src/main/java/com/f4/fqs/gateway_domain/domain/GatewayConstant.java
@@ -1,0 +1,5 @@
+package com.f4.fqs.gateway_domain.domain;
+
+public interface GatewayConstant {
+    String ROUTE_KEY_PREFIX = "route:"; // 키 접두사
+}
