+++ service/gateway/gateway_domain/src/test/java/com/f4/fqs/gateway_domain/GatewayDomainApplicationTests.java
@@ -1,0 +1,13 @@
+package com.f4.fqs.gateway_domain;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class GatewayDomainApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
