+++ service/gateway/gateway_server/src/main/java/com/f4/fqs/gateway/config/RedisConfig.java
@@ -1,0 +1,58 @@
+package com.f4.fqs.gateway.config;
+
+import com.f4.fqs.gateway_domain.domain.ApiRoute;
+import com.fasterxml.jackson.databind.DeserializationFeature;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.SerializationFeature;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.boot.context.event.ApplicationReadyEvent;
+import org.springframework.context.ApplicationListener;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Lazy;
+import org.springframework.data.redis.connection.ReactiveRedisConnectionFactory;
+import org.springframework.data.redis.core.ReactiveRedisTemplate;
+import org.springframework.data.redis.serializer.*;
+
+@Slf4j
+@RequiredArgsConstructor
+@Configuration
+public class RedisConfig implements ApplicationListener<ApplicationReadyEvent> {
+
+    private final ReactiveRedisTemplate<String, String> reactiveRedisTemplate;
+
+    @Override
+    public void onApplicationEvent(ApplicationReadyEvent event) {
+        reactiveRedisTemplate.opsForValue().get("1")
+                .doOnSuccess(i -> log.info("Initialize to redis connection."))
+                .doOnError((err) -> log.error("Failed to initialize redis connection: {}", err.getMessage()))
+                .subscribe();
+    }
+
+    @Override
+    public boolean supportsAsyncExecution() {
+        return ApplicationListener.super.supportsAsyncExecution();
+    }
+
+    @Bean
+    public ReactiveRedisTemplate<String, ApiRoute> reactiveRedisTemplate(@Lazy ReactiveRedisConnectionFactory connectionFactory) {
+        var objectMapper = new ObjectMapper()
+                .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
+                .registerModule(new JavaTimeModule())
+                .disable(SerializationFeature.WRITE_DATE_KEYS_AS_TIMESTAMPS);
+
+        Jackson2JsonRedisSerializer<ApiRoute> jsonSerializer = new Jackson2JsonRedisSerializer<>(objectMapper, ApiRoute.class);
+
+        RedisSerializationContext<String, ApiRoute> serializationContext = RedisSerializationContext
+                .<String, ApiRoute>newSerializationContext()
+                .key(RedisSerializer.string())
+                .value(jsonSerializer)
+                .hashKey(RedisSerializer.string())
+                .hashValue(jsonSerializer)
+                .build();
+
+        return new ReactiveRedisTemplate<>(connectionFactory, serializationContext);
+    }
+}
