+++ service/gateway/gateway_server/src/main/java/com/f4/fqs/gateway/infrastructure/repository/RouteRepository.java
@@ -1,0 +1,6 @@
+package com.f4.fqs.gateway.infrastructure.repository;
+
+
+//@Repository
+//public interface RouteRepository extends ReactiveCrudRepository<ApiRoute, String> {
+//}
