+++ service/gateway/gateway_server/src/main/resources/application.yml
@@ -1,0 +1,45 @@
+server:
+  port: 19091
+
+spring:
+  main:
+    web-application-type: reactive
+  application:
+    name: gateway-service
+  data:
+    redis:
+      host: localhost
+      port: 6379
+#      username: default
+#      password: systempass
+  cloud:
+    gateway:
+      routes:
+        - id: auth-service
+          uri: lb://auth-service
+          predicates:
+            - Path=/api/auth/**
+        - id: user-service
+          uri: lb://user-service
+          predicates:
+            - Path=/api/user/**
+        - id: queue-service
+          uri: lb://queue-service
+          predicates:
+            - Path=/api/queue/**
+        - id: queueManage-service
+          uri: lb://queueManage-service
+          predicates:
+            - Path=/api/queueManage/**
+      discovery:
+        locator:
+          enabled: true
+
+service:
+  jwt:
+    secret-key: "WnN3Q2FjQzRrNFl0V1d1TWp0RE5Xemw2RkE0WVYzS2VsR2xpUVF6M0x6dFNXbnc="
+
+eureka:
+  client:
+    service-url:
+      defaultZone: http://localhost:19090/eureka/
