--- service/gateway/src/main/java/com/f4/fqs/gateway/application/service/impl/ApiRouteLocatorImpl.java
@@ -1,53 +1,0 @@
-package com.f4.fqs.gateway.application.service.impl;
-
-import com.f4.fqs.gateway.application.service.RouteService;
-import com.f4.fqs.gateway.domain.ApiRoute;
-import io.micrometer.common.util.StringUtils;
-import lombok.RequiredArgsConstructor;
-import org.springframework.cloud.gateway.route.Route;
-import org.springframework.cloud.gateway.route.RouteLocator;
-import org.springframework.cloud.gateway.route.builder.BooleanSpec;
-import org.springframework.cloud.gateway.route.builder.Buildable;
-import org.springframework.cloud.gateway.route.builder.PredicateSpec;
-import org.springframework.cloud.gateway.route.builder.RouteLocatorBuilder;
-import org.springframework.stereotype.Service;
-import reactor.core.publisher.Flux;
-
-import java.util.List;
-import java.util.Map;
-
-@RequiredArgsConstructor
-@Service
-public class ApiRouteLocatorImpl implements RouteLocator {
-    private final RouteLocatorBuilder routeLocatorBuilder;
-    private final RouteService routeService;
-
-    @Override
-    public Flux<Route> getRoutes() {
-        RouteLocatorBuilder.Builder routesBuilder = routeLocatorBuilder.routes();
-
-        // List<ApiRoute>로 변경된 getAll() 메서드를 사용
-        List<ApiRoute> apiRoutes = routeService.getAll();
-
-        // apiRoutes를 Flux로 변환하여 map 연산 수행
-        return Flux.fromIterable(apiRoutes)
-                .map(apiRoute -> routesBuilder.route(String.valueOf(apiRoute.getRouteIdentifier()),
-                        predicateSpec -> setPredicateSpec(apiRoute, predicateSpec)))
-                .collectList()
-                .flatMapMany(builders -> routesBuilder.build().getRoutes());
-    }
-
-    private Buildable<Route> setPredicateSpec(ApiRoute apiRoute, PredicateSpec predicateSpec) {
-        BooleanSpec booleanSpec = predicateSpec.path(apiRoute.getPath());
-        if (!StringUtils.isEmpty(apiRoute.getMethod())) {
-            booleanSpec.and()
-                    .method(apiRoute.getMethod());
-        }
-        return booleanSpec.uri(apiRoute.getUri());
-    }
-
-    @Override
-    public Flux<Route> getRoutesByMetadata(Map<String, Object> metadata) {
-        return RouteLocator.super.getRoutesByMetadata(metadata);
-    }
-}
