--- service/gateway/src/main/java/com/f4/fqs/gateway/domain/ApiRoute.java
@@ -1,18 +1,0 @@
-package com.f4.fqs.gateway.domain;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.redis.core.RedisHash;
-
-@Getter
-@Setter
-@RedisHash("apiRoutes")
-public class ApiRoute {
-    @Id
-    private String id;
-    private String routeIdentifier;
-    private String uri;
-    private String method;
-    private String path;
-}
