--- service/gateway/src/main/java/com/f4/fqs/gateway/router/ApiRouteRouter.java
@@ -1,24 +1,0 @@
-package com.f4.fqs.gateway.router;
-
-import com.f4.fqs.gateway.handler.ApiRouteHandler;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.MediaType;
-import org.springframework.web.reactive.function.server.RouterFunction;
-import org.springframework.web.reactive.function.server.RouterFunctions;
-import org.springframework.web.reactive.function.server.ServerResponse;
-
-import static org.springframework.web.reactive.function.server.RequestPredicates.*;
-
-@Configuration
-public class ApiRouteRouter {
-    @Bean
-    public RouterFunction<ServerResponse> route(ApiRouteHandler apiRouteHandler) {
-        return RouterFunctions.route(POST("/routes")
-                        .and(accept(MediaType.APPLICATION_JSON)), apiRouteHandler::create)
-                .andRoute(GET("/routes/:routeId")
-                        .and(accept(MediaType.APPLICATION_JSON)), apiRouteHandler::getById)
-                .andRoute(GET("/routes/refresh-routes")
-                        .and(accept(MediaType.APPLICATION_JSON)), apiRouteHandler::refreshRoutes);
-    }
-}
