+++ settings.gradle
@@ -5,7 +5,8 @@ plugins {
 rootProject.name = 'fqs'
 
 include 'service:eureka'
-include 'service:gateway'
+include 'service:gateway:gateway_domain'
+include 'service:gateway:gateway_server'
 include 'service:auth'
 include 'service:user'
 include 'service:queue'
